ielts2_bot/requirements.txt
aiogram>=3.3.0
asyncpg>=0.29.0
google-genai>=1.0.0
gtts>=2.5.0
pydub>=0.25.1
python-dotenv>=1.0.0
